Fixing time unit.

+++ src/test/java/commons/cloud/ProviderTest.java
@@ -13,7 +13,7 @@ import commons.util.Triple;
 public class ProviderTest {
 	
 	private String name = "prov";
-	private static final int HOUR_IN_MILLIS = 1000 * 60 * 60;
+	private static final long HOUR_IN_MILLIS = 1000 * 60 * 60;
 	private JEEventScheduler scheduler;
 	
 	@Before
@@ -232,39 +232,39 @@ public class ProviderTest {
 		
 		//Adding reserved resources
 		Long machine1ID = 1l;
-		Triple<Double, Double, Double> triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 1d * HOUR_IN_MILLIS;
-		triple.secondValue = 5d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 1 * HOUR_IN_MILLIS;
+		triple.secondValue = 5 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine1ID, triple);
 		
 		Long machine2ID = 2l;
-		Triple<Double, Double, Double> triple2 = new Triple<Double, Double, Double>();
-		triple2.firstValue = 0d;
-		triple2.secondValue = 5d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple2 = new Triple<Long, Long, Double>();
+		triple2.firstValue = 0l;
+		triple2.secondValue = 5 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine2ID, triple2);
 		
 		Long machine3ID = 3l;
-		Triple<Double, Double, Double> triple3 = new Triple<Double, Double, Double>();
-		triple3.firstValue = 3d * HOUR_IN_MILLIS;
-		triple3.secondValue = 18d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple3 = new Triple<Long, Long, Double>();
+		triple3.firstValue = 3 * HOUR_IN_MILLIS;
+		triple3.secondValue = 18 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine3ID, triple3);
 		
 		Long machine4ID = 4l;
-		Triple<Double, Double, Double> triple4 = new Triple<Double, Double, Double>();
-		triple4.firstValue = 0d;
-		triple4.secondValue = 15d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple4 = new Triple<Long, Long, Double>();
+		triple4.firstValue = 0l;
+		triple4.secondValue = 15 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine4ID, triple4);
 		
 		Long machine5ID = 5l;
-		Triple<Double, Double, Double> triple5 = new Triple<Double, Double, Double>();
-		triple5.firstValue = 5d * HOUR_IN_MILLIS;
-		triple5.secondValue = 17.5d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple5 = new Triple<Long, Long, Double>();
+		triple5.firstValue = 5 * HOUR_IN_MILLIS;
+		triple5.secondValue = (long)(17.5d * HOUR_IN_MILLIS);
 		provider.reservedResources.put(machine5ID, triple5);
 		
 		Long machine6ID = 6l;
-		Triple<Double, Double, Double> triple6 = new Triple<Double, Double, Double>();
-		triple6.firstValue = 0.5d * HOUR_IN_MILLIS;
-		triple6.secondValue = 15.73d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple6 = new Triple<Long, Long, Double>();
+		triple6.firstValue = (long)(0.5d * HOUR_IN_MILLIS);
+		triple6.secondValue = (long)(15.73d * HOUR_IN_MILLIS);
 		provider.reservedResources.put(machine6ID, triple6);
 		
 		assertEquals(600 + 68 * reservedCpuCost + 68 * monitoringCost, provider.calculateCost(0), 0.0d);
@@ -290,9 +290,9 @@ public class ProviderTest {
 		
 		//Adding reserved resources
 		Long machine1ID = 1000l;
-		Triple<Double, Double, Double> triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 0d;
-		triple.secondValue = -5d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 0l;
+		triple.secondValue = (long)(-5d * HOUR_IN_MILLIS);
 		provider.reservedResources.put(machine1ID, triple);
 		
 		try{
@@ -322,33 +322,33 @@ public class ProviderTest {
 		
 		//Adding reserved resources
 		Long machine1ID = 1l;
-		Triple<Double, Double, Double> triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 0d;
-		triple.secondValue = 2d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 0l;
+		triple.secondValue = 2 * HOUR_IN_MILLIS;
 		provider.onDemandResources.put(machine1ID, triple);
 		
 		Long machine2ID = 2l;
-		Triple<Double, Double, Double> triple2 = new Triple<Double, Double, Double>();
-		triple2.firstValue = 1d * HOUR_IN_MILLIS;
-		triple2.secondValue = 2d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple2 = new Triple<Long, Long, Double>();
+		triple2.firstValue = 1 * HOUR_IN_MILLIS;
+		triple2.secondValue = 2 * HOUR_IN_MILLIS;
 		provider.onDemandResources.put(machine2ID, triple2);
 		
 		Long machine3ID = 3l;
-		Triple<Double, Double, Double> triple3 = new Triple<Double, Double, Double>();
-		triple3.firstValue = 0d;
-		triple3.secondValue = 15d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple3 = new Triple<Long, Long, Double>();
+		triple3.firstValue = 0l;
+		triple3.secondValue = 15 * HOUR_IN_MILLIS;
 		provider.onDemandResources.put(machine3ID, triple3);
 		
 		Long machine4ID = 450l;
-		Triple<Double, Double, Double> triple4 = new Triple<Double, Double, Double>();
-		triple4.firstValue = 1d * HOUR_IN_MILLIS;
-		triple4.secondValue = 2.2d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple4 = new Triple<Long, Long, Double>();
+		triple4.firstValue = 1 * HOUR_IN_MILLIS;
+		triple4.secondValue = (long)(2.2d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine4ID, triple4);
 		
 		Long machine5ID = 5l;
-		Triple<Double, Double, Double> triple5 = new Triple<Double, Double, Double>();
-		triple5.firstValue = 2d * HOUR_IN_MILLIS;
-		triple5.secondValue = 7.14d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple5 = new Triple<Long, Long, Double>();
+		triple5.firstValue = 2 * HOUR_IN_MILLIS;
+		triple5.secondValue = (long)(7.14d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine5ID, triple5);
 		
 		assertEquals( 26 * onDemandCpuCost + 26 * monitoringCost, provider.calculateCost(0), 0.0d);
@@ -374,9 +374,9 @@ public class ProviderTest {
 		
 		//Adding reserved resources
 		Long machine1ID = 1l;
-		Triple<Double, Double, Double> triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 0d;
-		triple.secondValue = -6d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 0l;
+		triple.secondValue = (long)(-6d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine1ID, triple);
 		
 		try{
@@ -386,9 +386,9 @@ public class ProviderTest {
 		}
 		
 		//FIXME: Double value explodes with large value!
-		triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 1000d * HOUR_IN_MILLIS;
-		triple.secondValue = 400d * HOUR_IN_MILLIS;
+		triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 1000 * HOUR_IN_MILLIS;
+		triple.secondValue = 400 * HOUR_IN_MILLIS;
 		provider.onDemandResources.clear();
 		provider.onDemandResources.put(machine1ID, triple);
 		
@@ -419,69 +419,69 @@ public class ProviderTest {
 		
 		//Adding reserved resources
 		Long machine1ID = 1l;
-		Triple<Double, Double, Double> triple = new Triple<Double, Double, Double>();
-		triple.firstValue = 0d;
-		triple.secondValue = 10d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple = new Triple<Long, Long, Double>();
+		triple.firstValue = 0l;
+		triple.secondValue = 10 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine1ID, triple);
 		
 		Long machine2ID = 2l;
-		Triple<Double, Double, Double> triple2 = new Triple<Double, Double, Double>();
-		triple2.firstValue = 10d * HOUR_IN_MILLIS;
-		triple2.secondValue = 30d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple2 = new Triple<Long, Long, Double>();
+		triple2.firstValue = 10 * HOUR_IN_MILLIS;
+		triple2.secondValue = 30 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine2ID, triple2);
 		
 		Long machine3ID = 3l;
-		Triple<Double, Double, Double> triple3 = new Triple<Double, Double, Double>();
-		triple3.firstValue = 1d * HOUR_IN_MILLIS;
-		triple3.secondValue = 16d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple3 = new Triple<Long, Long, Double>();
+		triple3.firstValue = 1 * HOUR_IN_MILLIS;
+		triple3.secondValue = 16 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine3ID, triple3);
 		
 		Long machine4ID = 4l;
-		Triple<Double, Double, Double> triple4 = new Triple<Double, Double, Double>();
-		triple4.firstValue = 50d * HOUR_IN_MILLIS;
-		triple4.secondValue = 65d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple4 = new Triple<Long, Long, Double>();
+		triple4.firstValue = 50 * HOUR_IN_MILLIS;
+		triple4.secondValue = 65 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine4ID, triple4);
 		
 		Long machine5ID = 5l;
-		Triple<Double, Double, Double> triple5 = new Triple<Double, Double, Double>();
-		triple5.firstValue = 0d;
-		triple5.secondValue = 15d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple5 = new Triple<Long, Long, Double>();
+		triple5.firstValue = 0l;
+		triple5.secondValue = 15 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine5ID, triple5);
 		
 		Long machine6ID = 6l;
-		Triple<Double, Double, Double> triple6 = new Triple<Double, Double, Double>();
-		triple6.firstValue = 11.5d * HOUR_IN_MILLIS;
-		triple6.secondValue = 24d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple6 = new Triple<Long, Long, Double>();
+		triple6.firstValue = (long)(11.5d * HOUR_IN_MILLIS);
+		triple6.secondValue = 24 * HOUR_IN_MILLIS;
 		provider.reservedResources.put(machine6ID, triple6);
 		
 		Long machine7ID = 7l;
-		Triple<Double, Double, Double> triple7 = new Triple<Double, Double, Double>();
-		triple7.firstValue = 0d;
-		triple7.secondValue = 15.23d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple7 = new Triple<Long, Long, Double>();
+		triple7.firstValue = 0l;
+		triple7.secondValue = (long)(15.23d * HOUR_IN_MILLIS);
 		provider.reservedResources.put(machine7ID, triple7);
 		
 		Long machine8ID = 8l;
-		Triple<Double, Double, Double> triple8 = new Triple<Double, Double, Double>();
-		triple8.firstValue = 18d * HOUR_IN_MILLIS;
-		triple8.secondValue = 36d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple8 = new Triple<Long, Long, Double>();
+		triple8.firstValue = 18 * HOUR_IN_MILLIS;
+		triple8.secondValue = 36 * HOUR_IN_MILLIS;
 		provider.onDemandResources.put(machine8ID, triple8);
 		
 		Long machine9ID = 9l;
-		Triple<Double, Double, Double> triple9 = new Triple<Double, Double, Double>();
-		triple9.firstValue = 0d;
-		triple9.secondValue = 78.5d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple9 = new Triple<Long, Long, Double>();
+		triple9.firstValue = 0l;
+		triple9.secondValue = (long)(78.5d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine9ID, triple9);
 		
 		Long machine10ID = 10l;
-		Triple<Double, Double, Double> triple10 = new Triple<Double, Double, Double>();
-		triple10.firstValue = 2.4d * HOUR_IN_MILLIS;
-		triple10.secondValue = 3.6d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple10 = new Triple<Long, Long, Double>();
+		triple10.firstValue = (long)(2.4d * HOUR_IN_MILLIS);
+		triple10.secondValue = (long)(3.6d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine10ID, triple10);
 		
 		Long machine11ID = 11l;
-		Triple<Double, Double, Double> triple11 = new Triple<Double, Double, Double>();
-		triple11.firstValue = 0d;
-		triple11.secondValue = 5.14d * HOUR_IN_MILLIS;
+		Triple<Long, Long, Double> triple11 = new Triple<Long, Long, Double>();
+		triple11.firstValue = 0l;
+		triple11.secondValue = (long)(5.14d * HOUR_IN_MILLIS);
 		provider.onDemandResources.put(machine11ID, triple11);
 		
 		assertEquals( 7 * reservationOneYearFee + 104 * reservedCpuCost + 104 * monitoringCost +
